chore(packages): simplify crypto-frontmatter
